requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
colorama==0.4.4
discord.py==1.7.2
Flask==2.0.0
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
gunicorn==20.1.0
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
multidict==5.1.0
psycopg2-binary==2.9.7
pycodestyle==2.7.0
pycparser==2.20
pyjokes==0.6.0
python-dotenv==0.17.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.15
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
Werkzeug==2.0.0
WTForms==2.3.3
yarl==1.6.3
